% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.curves.R
\name{fit.curves}
\alias{fit.curves}
\title{fit.curves}
\usage{
fit.curves(
  X,
  Y,
  Fname = "",
  Figname = "",
  FigDir = "",
  Ylab = "Y",
  Xlab = "X",
  axis.font = 0.75,
  log.base = NA,
  log.factor = 0,
  FigW = 7,
  FigH = 7,
  SigInit = c(0, 0),
  Uni6Init = c(0, -1, 2, 2, 2, 2),
  Uni5Init = NA,
  reltol = 1e-04
)
}
\arguments{
\item{X}{concentrations}

\item{Y}{response values}

\item{Fname}{filename for saving figures}

\item{Figname}{Title to appear on figures}

\item{FigDir}{Directory where figures will be saved}

\item{Ylab}{Ylabel for figures}

\item{Xlab}{Xlabel for figures}

\item{axis.font}{fontsize for figures}

\item{log.base}{What base used to transform concentrations (NA = no transformation)}

\item{log.factor}{value added to log-transformed data to avoid log(0)}

\item{FigW}{Figure width}

\item{FigH}{Figure height}

\item{SigInit}{2-element vector with start values for sigmoidal function}

\item{Uni6Init}{6-element vector with start values for 5-param unimodal}

\item{Uni5Init}{5-element vector with start values for 6-param unimodal}

\item{reltol}{numerical tolerance for mle2}
}
\value{
baseline mle2 object for null model

sigmoidal mle2 object for sigmoid DR curve

linear mle2 object for linear DR curve

unimodal5 mle2 object for 5-param unimodal DR curve

unimodal6 mle2 object for 6-param unimodal DR curve

quadratic mle2 object for quadratic DR curve

AIC.table Table with AICc results for each model

ANOVA Table with likelihood ratio test results for each model

FN File name used for plotting

B Identity of 'best' model as judged by likelihood ratio test

Pval Table of likelihood ratio test p-values
}
\description{
Fit dose-response curves using MLE
}
