% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmoidDR.R
\name{SigmoidDR}
\alias{SigmoidDR}
\title{SigmoidDR}
\usage{
SigmoidDR(p, X, Y)
}
\arguments{
\item{p}{1x5 vector of parameters: g = max Y value, h = min Y value, a = exponential intercept term, b = exponential rate term, and s = standard deviation}

\item{Y}{concentration values for each observation}

\item{C}{response value for each observation (not used)}
}
\value{
negative log-likelihood of the model fit
}
\description{
4-parameter sigmoidal dose-response curve, for MLE estimation. Used as an objective function for mle2
}
